Add set_current_process function, never set CURRENT_PROCESS directly
Setting CURRENT_PROCESS directly causes a bit of a recursiveness issue: The
Drop is called before changing CURRENT_PROCESS. If the Drop calls
get_current_process, it will revive the Arc... which causes all kinds of
weirdness.

Furthermore, it gives us a convenient place where we can set the pstate to
Running. Note that we only change it to Running if it was in Scheduling state,
since we might have other states that have "higher priority" (Killed state,
comming in a future commit)