PhysicalMemRegion and FrameAllocator
Deleted the struct Frame, now the kernel works only on
PhysicalMemRegion, which represents a range of consecutive
physical frames.

The FrameAllocator was completely re-written to support this api,
and now gives back PhysicalMemRegion when called for allocation.
When this struct is dropped, the frames are automatically given
back to the FrameAllocator, making them available again.

The FrameAllocator also supports fragmented allocation, by
returning a Vec<PhysicalMemRegion>.

The idea of what a FrameAllocator should be able to do was generalized
in an arch-independent FrameAllocatorTrait,
and arch-specific implementations only need to implement this trait.