Userspace Bookkeeping
The UserLand of every process is bookkeeped, with all of its mappings tracked.

We store all the mappings of a process in a BTreeMap with the virtual address as the key,
and implement a few functions to add, remove, and retrieve a mapping.
We also provide a function to find the closest mapping of a given address,
this will be used by the QueryMemory syscall.

The bookkepping is in charge of tracking the physical memory pointed to by a mapping.
When the mapping is removed, those frames are automaticly freed.
The page tables should now be simply a reflection of the bookkeeping state.

We now support sharing a mapping between multiple processes.
A shared mapping is stored in the bookkeeping as an Arc<Vec<PhysicalMemRegion>>,
this way the frames are freed only when they are no longer mapped in any process.

We also have support for splitting a mapping, although it is not currently implemented in the api.