Better CrossProcessMapping
mirror_mapping now takes a length as argument (the length *must* be within a
single mapping).
CrossProcessMapping now has an offset and a length. The offset is used to return
the KernelAddress of the area the user asked to remap (which may be different
from the first area of Mapping).