Introducing threads
We now run threads.
A thread has a link to the process it belongs to.
A process has an array of links to the threads it owns.

When its last thread is dropped, a process is also dropped, automatically.
This is done with clever reference counting.

This patch splits the ProcessStruct in parts that are thread-specific
(kstack, hwcontext, ...) and belong in the ThreadStruct, and parts
that are process-shared (pid, pmemory, name), and stay in the ProcessStruct.

Changed scheduler to work on threads instead of processes.