Thread creation in 1 step
Now that ThreadStruct creation is done after ProcessStruct creation,
we don't need the NotReady -> Readying -> Stopped state cycle
done by set_start_arguments.

Instead, a thread is created directly with its pc and sp set, and put
in a Newborn state. When svcStartThread is called, it checks it
was still in a Newborn state, and adds it to the schedule queue.