frame allocator check _before_ instantiation
When recreating a mapping, we used to instantiate a PhysicalMemRegion,
and then check it was unallocated and panic otherwise.

But when this panics, the unwinder will attempt to drop the
invalid PhysicalMemRegion that we created, which will try to free it,
and panic again because is was not allocated in the first place.