Kernel capabilities: Implement syscall authorization
A process can only call an SVC that he asked for in its capabilities. Attempting
to call a forbidden SVC is treated the same way as calling an invalid SVC, which
now kills the process. We'll eventually want to implement usermode exception
handling, but that's left for a future PR.

In order to avoid crashes, we now use .get when getting the SYSCALL_NAME.