Add argument handling to start_thread syscall.
Re-enable the test_threads command from shell.
Some shenanigans were used to make it all work. The terminal is
passed by value to the test_threads function, which wraps it in
an Arc<Mutex> in order to share it between the two threads. Those
threads will then repeatedly try to lock it, and yield if they fail.
Yielding is important, since we lack preemption.

The argument is passed via EAX. This sucks because the C ABI expects
it on the stack, so we have to make a wrapper naked function that
pushes it... Maybe it should be the job of the kernel? On the other
hand, we're going to want a wrapper around test creation anyways.