update rust-toolchain
nightly-2018-12-31 no longer compiles, compiler_builtins
was updated and now fails, and as it is not tracked by Cargo.lock,
it was not pinned to a specific version.

I just picked dates at random and this is the first one for which
core managed to compile successfully.

fixes #194

Since 1.33 const_let has been stabilized, so we no longer need the
feature attribute.

Also getting some warnings on non-uppercase globals.