ahci driver
Introducing the driver for AHCI disks, for ATA and SATA devices.

Provides IPC endpoints to read and write some sectors.

For now it is a very modest single-threaded driver,
blocking on every request until the previous one is completed,
but this is prone to change as AHCI supports up to 32 simultaneous
outstanding commands.