vi: BACKBUFFER_ARR resolution
The backbuffer is no longer hardcoded for 1280x800,
but instead holds enough memory for almost all screen sizes.

I could have dynamically allocated it, but I liked how
it resides in the .bss, and wanted to keep it that way.

Fixes #79

This revealed a problem with .bss and our userspace.ld, see #216