shell: test_page_fault asm
Rewrote the test_page_fault function in asm, because rust
used to optimize it out since it is extremely UB.

Fixes #223

I went and checked in GDB, and @roblabla was right, the match branch
was never taken, but instead we ended up in weird places.

Now it seems to be working... well, crashing ... well, doing what we expect.