Libuser: Use u64 instead of usize for the address
We do lots of shifting operations on the address and sizes.
Unfortunately, it's come to my attention that x1 >> x2 has a completely
broken behavior if x2 is higher than the bitwidth of x1 (instead of
returning the obvious 0). This breaks the implementation on 32-bit.

To remediate this, we simply move all the addresses and sizes to u64.