SwIPC generator: the awakening
Implement an IPC client code generator based on SwIPC definitions.
The goal is to eventually remove all manually defined IPC client code
from libuser in favor of being all automatically generated.

This first draft defines vi.id using SwIPC, and swaps the libuser manual
client code with this version. The generator is implemented as a proc
macro. The parser is implemented using a pest grammar and a quick-n-dirty
AST layered on top.