Renaming KFS to SunriseOS
See #197

The crates are now:

* sunrise-kernel
* sunrise-bootstrap
* sunrise-shell
* sunrise-libuser
* sunrise-clock
* sunrise-sm
* sunrise-vi
* sunrise-ahci
* sunrise-libutils
* sunrise-libkern

Only the README.md and the welcome message at boot use the full "SunriseOS" name.
Every other places use the shorter name "Sunrise" if possible, or "sunrise"
when lowercase is preferred.

For now all components are named after the (short) name of the project, i.e. the kernel
is "the Sunrise kernel", sm is "the Sunrise service manager", etc.