Fix documentation building.
The introduction of SwIPC-gen in the workspace broke documentation
building due to it depending on stdlib, which is not available when
building with --target=i386-unknown-none.

Documentation shall now be built without passing a target. Instead, any
cfg blocking out based on target architecture or OS should additionally
allow rustdoc, e.g:

```rust
```

This will allow rustdoc to punch through the cfg. It brings a few issues
however: we cannot define lang items, as they would be duplicated.
Those should not allow rustdoc through. Furthermore, we might have
problem in the future with multiple architectures. We'll see when it
comes.