SwIPC Gen: Use hard to guess variable names in fn
We currently use common variable names in the function implementation,
such as buf, msg, etc... This can cause conflicts with argument names,
causing hard to debug errors.

Ideally, we'd create separate spans for everything, for proper hygiene.
But doing this would require refactoring the code to not use strings
everywhere. As a temporary band-aid, we should suffix __ to all internal
variable names.