Kernel: CrossProcessMapping holds a Mapping
CPM now holds a Mapping instead of underlying Arc, and delegates most of
the offset/length logic to the Mapping::new constructor. This simplifies
the implementation of CPM a whole lot.

Add a Mapping::frames_it() function that returns an iterator on the
frames it maps. It takes into account the phys_offset/length so that it
only returns the frames actually mapped.