Kernel: Use IOAPIC and LAPIC instead of the PIC
PIC is so 1980s, IOAPIC is so 200s.

Moving to the APIC provides us with two immediate and one long term
benefit:

1. It allows us to use MSI for PCI interruptions, which should cleanly
   resolve the problem that level triggered interrupts don't work well
   in the context of a microkernel.

2. It provides us with a nice speed boost compared to the PIC, what with
   the IO-APIC being a lot cleaner to use.

3. It will eventually allow SMP.

The IO-APIC is mostly wired in the same way the PIC is. However, under
QEmu, there's a bug that prevents using IRQ line 0, so we have to move
the timer to another IRQ line (we use 2 since that's the recommended
line for timers).