Libuser: Start setting up object macro replacement
The object macro has outlived its usefulness: it's hard to maintain and
gives sub-par error messages. We're going for a replacement based on
swipc-gen, generating traits for every IPC interface.

Problem is, we can't use the Object trait anymore either. Because of the
coherence rules, we can't make multiple blanket implementation of
Object for each IPC interface. So instead, SessionWrapper and
PortHandler will now take a dispatch function directly as an argument,
and won't rely on the Object macro directly. This dispatch function will
be generated automatically by the swipc-gen.