Kernel: fix paging table leak
Fixes #313, see ticket for context.

Removed the `destroy()` function of InactiveHierarchyTrait,
as it was never called, and also invalid (used to free the KernelLand tables,
which it shouldn't)

Now the arch implementation of paging is in charge of freeing the tables,
and any additional memory relative to the hierarchy (none for x86).