Kernel: thread argument ABI + main thread handle
Passing its thread handle to the first thread of a process,
and changing the ABI for passing arguments to a thread.

The argument passed to a thread used to be stored in its ThreadStruct at thread creation.

First there was no real reason for it to be kept in the ThreadStruct for all of the thread's lifetime,
And secondly it forced us to already know the argument at the time we create the ThreadStruct,
which is not the case for main threads, the argument being a handle to the ThreadStruct we haven't yet created.

This lead to a redesign of the way a thread gets its argument.

The argument is now passed on the kernel stack of a new thread, just like the entrypoint
and userspace stack address.

This is done by `prepare_for_first_schedule`, which is called much after the ThreadStruct creation,
so it gives us more time to figure out the argument.

When a thread is first scheduled, it starts in `first_schedule`, with the argument in one of its registers.
It backups it just like the entry_point, finishes the schedule-in, and jumps to userspace with it.