Shell: fix test_threads
The `test_threads` command used to display a seemingly random list
of variable number of As and Bs.
This was due to the fact that the `write!` macro internally makes
IPC calls, which reschedules the thread, while it is still holding
the lock.

When the concurrent threads gets to run, it sees that the lock is
held, so it does nothing and skips its turn silently.

We now make sure that every thread prints exactly 10 times,
not counting the rounds that are skipped.

This closes #224