Kernel: implement get_system_tick.
The GetSystemTick syscall returns a monotonic up-counter representing a
frequency of 19200000 Hz. Note that none of our clocks actually run at
this frequency (HPET is likely to have a much higher frequency, while
PIT is likely to have a much, much, much lower frequency). As such, we
have to do some maths.

In order to implement GetSystemTick, the KERNEL_TIMER_INFO now contains
a get_tick function pointer. When the kernel timer gets selected, it
should provide an implementation of that function as a parameter to
set_kernel_timer_info.