Kernel: use 2 TLS segments
Found out the hard way that we cannot use the TLS region as an elf tls TCB by tricking the compiler and putting a pointer to our actual TCB in the first word of the TLS region, because for the GNU model llvm will generate code such as

```asm
	mov eax, gs:0xfffffffc
```

which gets the thread local variable directly from an offset from `gs`, instead of the usual

```asm
	mov eax, gs:0
	mov eax, [eax - 0x4]
```

which would have worked.

This would have meant we should have had our elf thread local variables directly prior to the TLS region, which is allocated by the kernel, so this is impossible.

Instead we choose to use two thread-local segments:

* `fs` points to the TLS region. It is allocated by the kernel, and cannot be changed.
* `gs` points to the TLS elf. The address it points to is controlled by the userspace, through the set_thread_area syscall.

When a user thread starts, it is expected to allocate a memory region, copy its `.tdata` and `.tbss` to it, and call set_thread_area to make `gs` point to it.

The kernel is still in charge of swapping the `gs` and `fs` segments on every thread switch.