Kernel: panic takes a panic origin
Simplifying the panic function by taking a clear cause for the panic.

This way we can easily filter what's interesting to display, and what's not.
The caller of kernel panic shouldn't have to compute what to display himself.

Also adding Blue Screen Of Death so the user knows we have panicked.