Kernel: generate_trap_gate_handler! macro
* Context *

Exception handlers used to have no asm wrapper that saves the registers (userspace hardware context) and instead use the x86-interrupt calling convention to do the dirty work for us.
This had limitations, a x86-interrupt function does not expose the registers it has backed up. This is needed to implement a debugger so we can inspect the state of a thread's registers when it hits a breakpoint, and possibly modify them.

This meant that x86-interrupt exception had to eventually be replaced by a true asm wrapper.

* TLS switching *

For TLS + cpu-locals to work, because they both use `gs` to point to their thread_local area we must switch where `gs` points to every time we enter and leave the kernel. This happens at:

* Syscalls
* Exceptions
* Interrupts

This means we now need to add a small stub that switches `gs` at the start and end of every exception handler. Instead of doing that, we choose this time to finally implement an asm wrapper for exceptions.

This commit introduces the `generate_trap_gate_handler!` macro, which will generate an asm wrapper for exceptions, and handle all the boilerplate code and condition checking that exceptions have to perform, so that user-defined exception handlers cannot forget to perform them.

The `generate_trap_gate_handler!` macro is general enough for syscalls to use it. I'm planning to do exactly that, but in a separate commit so we can revert easily if things go wrong.

Interrupts handlers still use the x86-interrupt calling convention, as interrupt handling is supposed to be dead-simple, not make use of any cpu-locals, and return as fast as possible, we don't even want the overhead of saving the userspace hardware context.