Kernel: irqs use generate_trap_gate_handler!
Irqs used to follow the x86-interrupt calling convention.

Turns out I was wrong, irq handlers are not as simple as I thought, and we **do** access some cpu-local statics in irqs, so we need to switch `gs` everytime we enter the kernel.

We now treat them the same way as exceptions, since we need to switch `gs` everytime we handle one, and check if we're not killed when returning to userspace.

This adds the overhead of backuping registers to the ThreadStruct, however it might become useful when we will have a debugger, so we consider it tolerable.