IPC Gen: Add Proxy::new() returning shared handle
Introduce a Proxy::new() function, which caches the returned handle to
avoid creating new handles each time we need to access a service. Those
handle will be kept in a static Once, and never get dropped.

NOTE: We will want to eventually find a way to clean up the handles on
process exit.