Kernel: Implement ProcessStruct::check_range
This function mimicks nintendo's KMemoryManager::CheckRange. It goes
over a memory region and checks that it is in the expected state (both
in terms of MemoryState, MemoryPermissions and MemoryAttributes), and
that it is homogenous (that is, all submappings have the same state and
perms).

This is used to significantly simplify set_process_memory_permission and
map_process_memory.