Kernel: create_first_process locking
When creating the first ProcessMemory, we hold an &mut to it, and can call `get_mut` on its preemptive locks to get mut access without locking, instead of `lock` which would try to get the CURRENT_THREAD and panic, as we're in the process of creating it.

Also put `tls_elf` pointer behind a SpinLock instead of preempting lock, which seems more appropriate for a simple pointer.