Kernel: Fix random userspace gs = 0
Sometimes, when running userspace, GS would be 0, leading to
instructions accessing gs:[0] to fail with a null deref. I'm... still
not entirely sure why that happens. It appears that our ring detection
in trap_gate_asm fails, leading to gs not being changed before returning
to userspace. But what confuses me is that when this happens, gs should
be set to the kernel TLS, and not 0.

Anyways, this is fixed by pushing GS on the UserSpaceHardwareContext and
restoring it unconditionally before returning.