Kernel: Avoid UB in jump_to_entrypoint
jump_to_entrypoint could potentially generate invalid code due to not
providing enough information to the compiler about which registers it's
allowed to use.

We now enforce that arguments are passed to the correct register. This
might be used by the compiler to optimize its register scheduling, which
is cool too.

As an added defense-in-depth, init_main_thread, which is only called
from asm, takes an u32 and manually creates its ThreadHandle. This avoid
hard-to-debug problems due to Handle being a NonZeroU32.