Vi: Move terminal handling to vi
We now have a new IPC interface, twili::IPipe, representing a pipe on
which we can read and write data.

Vi now implements the font rendering and other associated logic with
terminal handling. It does so by implementing the IPipe interface.

We have lost the ability to print colors in the terminal. This will be
reinstated in a future commit, which will add ansi escape codes.