Fix deadlock.
This was interesting to figure out.

In short, when a scheduler based process switch occurs, there is a lock
held exclusively for enabling and disabling interrupts. This lock is not
dropped before the process switch.

If this is not the first time this process is run, it evens out by
dropping the copy of this lock held by the 'new' process. However, if it
is *not* the first time the process is run, there is no lock, and the
counter never gets decremented.

This commit fixes the issue by decrementing the counter in the new
process right before the iret which reenabled interrupts.